Svol0 bumper error if continued triggered (#2)
* Update config_example.h

In case of continuously triggering the bumper sensor, e.g. sensor is mechanically stuck, the mower only tries for a configurable time to avoid the supposed obstacle.
After the time is exceeded, the mower goes into error mode.
The time can be set in config.h by "BUMPER_MAX_TRIGGER_TIME". Unit is in seconds. A "0" (zero) disables the monitoring of the sensor.

* Update robot.cpp